Make heapsize trait public and move it into arrow-buffer
